requirements.txt
pydantic==1.9.1
typing==3.7.4.3
psycopg2-binary==2.9.3
pandas==1.4.2
